+++ plugin/httptrace/httptrace.go
@@ -29,8 +29,12 @@ var (
 	HostKey = key.New("http.host")
 	URLKey  = key.New("http.url")
 
-	propagator  = trace.DefaultHTTPPropagator()
-	propagators = propagation.New(propagation.WithInjectors(propagator), propagation.WithExtractors(propagator))
+	tcPropagator = trace.DefaultHTTPPropagator()
+	ccPropagator = correlation.DefaultHTTPPropagator()
+	propagators  = propagation.New(
+		propagation.WithInjectors(tcPropagator, ccPropagator),
+		propagation.WithExtractors(tcPropagator, ccPropagator),
+	)
 )
 
 // Returns the Attributes, Context Entries, and SpanContext that were encoded by Inject.
